[Task-2-1] Some refactoring
