--- vadim.nesmeianov/task-2-1/internal/room/room.go
@@ -1,2 +1,0 @@
-package ac
-
